FutureYou/bullhorn/placementsApprovalEventSubscription/requirements.txt
requests
dotenv
python-dotenv
google-api-python-client==2.149.0
google-auth==2.35.0
google-auth-oauthlib==1.2.1
